shared/autoperipherals/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "autoperipherals"
dependencies = ["pyxdg", "click", "pydantic"]
dynamic = ["version"]

[project.scripts]
autoperipherals = "autoperipherals.cli:main"

[tool.setuptools.packages.find]
where = ["."]
